add: finish routes of lab all

+++ medrouter-api/src/docs/docs.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, BadRequestException } from '@nestjs/common';
 import { v4 as uuidv4 } from 'uuid';
 import * as fs from 'fs';
 import * as PdfPrinter from 'pdfmake';
@@ -6,6 +6,7 @@ import * as PdfPrinter from 'pdfmake';
 import { DocRepository } from './doc.repository';
 import { DocDto } from './dto/doc.dto';
 import { Doc } from './models/doc.entity';
+import { ExamStatus } from 'src/exams/enums/status.enum';
 
 @Injectable()
 export class DocsService {
@@ -23,6 +24,23 @@ export class DocsService {
   }
 
   async delete(id: any): Promise<any> {
+    const query = this.docRepository.createQueryBuilder('documents');
+
+    query.andWhere('documents.id = :id', { id });
+
+    query.andWhere('exam.status = :status OR exam.status = :status2', {
+      status: ExamStatus.CONCLUDED,
+      status2: ExamStatus.AVAILABLE,
+    });
+
+    const file = await query
+      .leftJoinAndSelect('documents.exam', 'exam')
+      .getOne();
+
+    if (file) {
+      throw new BadRequestException('File cannot be erased anymore');
+    }
+
     return this.docRepository.delete({ id });
   }
